% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pytha.R
\name{pytha}
\alias{pytha}
\title{Pythagoras Theorem}
\usage{
pytha(a, b)
}
\arguments{
\item{a}{numeric positive integer}

\item{b}{numeric positive integer}
}
\description{
This function can compute the third value in any pythagorean set
The user must provide either a & b or a & c or b  c
}
\examples{
pytha(4,5)
}
\author{
Edem Defor
}
