% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trmean.R
\name{trmean}
\alias{trmean}
\title{Trimmed Mean}
\usage{
trmean(x, s, l)
}
\arguments{
\item{x}{numeric vector}

\item{s}{numeric positive integer}

\item{l}{numberic positive integer}
}
\description{
This function can compute the trimmed mean
The user must provide a vector
}
\examples{
trmean(c(1:5),1,2)
}
\author{
Edem Defor
}
